+++ src/modules/status.js
@@ -1,0 +1,10 @@
+export const toggleTaskCompletion = (tasks, index) => {
+    tasks[index].completed = !tasks[index].completed;
+    localStorage.setItem('tasks', JSON.stringify(tasks));
+  };
+  
+  export const clearCompletedTasks = (tasks) => {
+    const updatedTasks = tasks.filter(task => !task.completed);
+    localStorage.setItem('tasks', JSON.stringify(updatedTasks));
+    return updatedTasks;
+  };
