+++ src/taskStatus.js
@@ -1,12 +1,12 @@
 
-export const toggleTaskCompletion = (tasks, index) => {
-    tasks[index].completed = !tasks[index].completed;
-    localStorage.setItem('tasks', JSON.stringify(tasks));
-  };
+// export const toggleTaskCompletion = (tasks, index) => {
+//     tasks[index].completed = !tasks[index].completed;
+//     localStorage.setItem('tasks', JSON.stringify(tasks));
+//   };
   
-  export const clearCompletedTasks = (tasks) => {
-    const updatedTasks = tasks.filter(task => !task.completed);
-    localStorage.setItem('tasks', JSON.stringify(updatedTasks));
-    return updatedTasks;
-  };
+//   export const clearCompletedTasks = (tasks) => {
+//     const updatedTasks = tasks.filter(task => !task.completed);
+//     localStorage.setItem('tasks', JSON.stringify(updatedTasks));
+//     return updatedTasks;
+//   };
